src/plugins/spheroid_plugin.py
from src.plugin_manager import OrganoidPlugin, register_plugin
import numpy as np
from typing import Dict, Any, List
from skimage import measure
import logging

logger = logging.getLogger(__name__)

@register_plugin("organoid", "spheroid")
class SpheroidPlugin(OrganoidPlugin):
    """球状类器官分析插件"""
    
    version = "1.0.0"
    
    @classmethod
    def get_required_configs(cls) -> List[str]:
        return ["size_range", "sphericity_threshold"]
    
    def define_morphology(self) -> Dict[str, Any]:
        return {
            'expected_shape': 'spherical',
            'size_range': self.config['size_range'],
            'sphericity_threshold': self.config['sphericity_threshold']
        }
        
    def analyze(self, image: np.ndarray) -> Dict[str, Any]:
        """分析球状类器官"""
        try:
            # 基本测量
            props = measure.regionprops(image.astype(int))[0]
            
            # 计算球形度
            volume = props.area
            surface_area = self._calculate_surface_area(image)
            sphericity = self._calculate_sphericity(volume, surface_area)
            
            # 检查是否满足形态要求
            min_size, max_size = self.config['size_range']
            is_valid = (min_size <= volume <= max_size and 
                       sphericity >= self.config['sphericity_threshold'])
            
            return {
                'diameter': self._calculate_diameter(props),
                'volume': volume,
                'surface_area': surface_area,
                'sphericity': sphericity,
                'is_valid_spheroid': is_valid,
                'centroid': props.centroid,
                'orientation': props.orientation
            }
            
        except Exception as e:
            logger.error(f"Error in spheroid analysis: {str(e)}")
            raise
            
    def _calculate_diameter(self, props) -> float:
        """计算等效直径"""
        return 2 * (3 * props.area / (4 * np.pi)) ** (1/3)
    
    def _calculate_surface_area(self, mask: np.ndarray) -> float:
        """计算表面积"""
        # 实现表面积计算逻辑
        pass
    
    def _calculate_sphericity(self, volume: float, surface_area: float) -> float:
        """计算球形度"""
        return ((6 * volume * np.pi) ** (2/3)) / surface_area 