requirements.txt
# 基础科学计算包
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# 图像处理
scikit-image>=0.18.0
Pillow>=8.2.0
opencv-python>=4.5.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 机器学习
scikit-learn>=0.24.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0

# 数据序列化
pyyaml>=5.4.0
json5>=0.9.5

# 图像分割模型
segment-anything>=1.0
transformers>=4.11.0  # for DINO

# 工具包
tqdm>=4.61.0  # 进度条
joblib>=1.0.0  # 并行计算
h5py>=3.3.0   # 大文件存储
pathlib>=1.0.1 # 路径处理

# 开发工具
pytest>=6.2.0  # 测试框架
black>=21.6b0  # 代码格式化
isort>=5.9.0   # import 排序
flake8>=3.9.0  # 代码检查 